Fix result type of index_put in slice+copy_ recomposition
The `copy_` op being replaced by `RecomposeSliceCopy_` operates on a
subset of the tensor being mutated, while the `index_put` op being
used to replace the `copy_` op operates on the entire tensor being
mutated. This means that the result type of the `index_put` should be
the type of the input to `index_put` and we need to make sure that
`copy_` does not have users before replacing to avoid type conflicts.

Note: this commit fixes a type conflict that only seems to arise when
`use_tracing=True`, since normally the recomposition happens before
type propagation takes place. Since the e2e testing framework does not
do tracing, here we use a lit test to check correctness.