upstream slice fixes (#35)
* Fix result type of index_put in slice+copy_ recomposition

The `copy_` op being replaced by `RecomposeSliceCopy_` operates on a
subset of the tensor being mutated, while the `index_put` op being
used to replace the `copy_` op operates on the entire tensor being
mutated. This means that the result type of the `index_put` should be
the type of the input to `index_put` and we need to make sure that
`copy_` does not have users before replacing to avoid type conflicts.

Note: this commit fixes a type conflict that only seems to arise when
`use_tracing=True`, since normally the recomposition happens before
type propagation takes place. Since the e2e testing framework does not
do tracing, here we use a lit test to check correctness.

* Add alias analysis for cast-like ops to maximize-value-semantics

When `use_tracing=True` is used to import a model into Torch-MLIR,
several casts get inserted in the IR to bridge the untyped inputs and
outputs with the typed body of the computation. These casts create
extra aliases of tensors that cause the current analysis in
`maximize-value-semantics` to fail.

In particular, the `maximize-value-semantics` analysis assumes that the
only valid alias right after an overwrite is the overwritten
alias. So, if there is a use of a casted version of the overwritten
alias after the overwrite, the analysis fails.

This commit improves the analysis by identifying all cast-like aliases
of the overwritten alias and allowing such aliases to be used after an
overwrite.

Because this issue only arises when using tracing, it cannot be
currently tested e2e, so only lit test is added.

* RecomposeComplexOps: Remove dead slice op

---------

Co-authored-by: Ramiro Leal-Cavazos <ramiroleal050@gmail.com>