fix(TorchToTosa.cpp): adjust torch div conversion
check the return type of the division to figure out whether to use
the floating point implementation of a division or to use the
integer.

the issue rose from the fact that the inputs are all integer but
the result was casted to floating point. The conversion then chose
to use the integer implementation of division which is not legal in
tosa when all the inputs get casted to floating point.